Finished wrapping up the firmware loading code.
I'm not sure whether the firmware has actually been loaded in successfully
or not.

fwcmd.h has the firmware APIs from carl9170.  It's GPLed (for now); but
it defines the API needed to talk to the firmware.

Next - implementing basic command send/response, so the echo test command
can be sent.