Continue fleshing out command queuing and completion.
This code (like the previous) is based off on what if_uath implements,
save that I'm trying to split it up into 'generic command queue handling'
and 'stuff to push into USB.'

So far it:

* allocates a command buffer
* queues the echo command
* starts a USB transfer
* sets up the USB receive response
* .. and prints that it received something; I haven't dug into -what-
  just yet.

Things I've learnt thus far:

* I'd like to split up attach() so it does a deferred firmware load,
  verify, load-into-NIC and echo request.  Doing it in probe seems
  .. dirty.

* But it does work; I get a response from the firmware; but no idea
  if it's valid or not yet.

* The command response comes in on the RX endpoint, not the IRQ
  endpoint.  Good to know.

* The linux code uses packed structures rather than Sam's MS/SM style
  field accessors from the ath/hal code.  I prefer the latter.

So far, so good.

I next need to check whethre the transfers need to be sized in DWORD
multiples; then implement that.