Convert match/attach and some of the USB worker function definitions over to the new world order.
Add an initial lock.

The USB transfers haven't yet been converted over to the new world order;
that'll come next.

otus_attachhook() and the firmware loader are the main bits that need to
come up in order to at least get the thing to probe/attach and respond
to a firmware query.  That's a good milestone.