Begin conversion of the HTC code over from Linux to BSD:
* debug flag naming
* debug method naming
* 'ar' -> 'sc'
* spin_lock -> mutex operations

It still doesn't completely compile, and for good reason.
So far it still assumes we're using skb's, etc.  We can't do that.
We have to convert the TX path over to take athp_buf's that we
setup and add a callback method to, and the RX path over to use
mbufs properly.  Or, athp_bufs; not sure yet.

Still, it's slowly getting there.