Prepare for some skb related sins.
ath10k assembles the contents of packets and headers by using the
skb hijinx functions (skb_reserve, skb_push, skb_put).  Now, we don't
currently have nice analogues to this in the freebsd mbuf representation -
we'd have to hand roll them ourselves.

Until I figure out a clean way to do it /without/ using len/size
in an athp_pbuf, let's track the original buffer size and current length
like we are currently doing.

When we allocate an mbuf m_len is set to 0, so in order to use it in
say the RX path, we'd need to know the original size of the mbuf.
We should "know" that via mbufs that use external storage with
m_ext.ext_size, but I don't know if that holds true for /all/ mbufs.
But we don't.  So, let's use fields in athp_buf until I'm sure
this will work fine, and then I can undo it.

At least this way I'm forced to use athp_bufs everywhere and methodize
this adjustment nonsense.

(But, I may give up and revert this commit anyway once I figure out
a slightly nicer way of doing things that don't involve athp_pbufs.)